+++ ovos_color_parser/version.py
@@ -1,0 +1,6 @@
+# START_VERSION_BLOCK
+VERSION_MAJOR = 0
+VERSION_MINOR = 0
+VERSION_BUILD = 1
+VERSION_ALPHA = 1
+# END_VERSION_BLOCK
