+++ setup.py
@@ -1,0 +1,80 @@
+import os
+
+from setuptools import setup
+
+BASEDIR = os.path.abspath(os.path.dirname(__file__))
+
+
+def package_files(directory):
+    paths = []
+    for (path, directories, filenames) in os.walk(directory):
+        for filename in filenames:
+            paths.append(os.path.join('..', path, filename))
+    return paths
+
+
+def required(requirements_file):
+    """ Read requirements file and remove comments and empty lines. """
+    with open(os.path.join(os.path.dirname(__file__), requirements_file),
+              'r') as f:
+        requirements = f.read().splitlines()
+        return [pkg for pkg in requirements
+                if pkg.strip() and not pkg.startswith("#")]
+
+
+def get_version():
+    """ Find the version of the package"""
+    version_file = os.path.join(BASEDIR, 'ovos_color_parser', 'version.py')
+    major, minor, build, alpha = (None, None, None, None)
+    with open(version_file) as f:
+        for line in f:
+            if 'VERSION_MAJOR' in line:
+                major = line.split('=')[1].strip()
+            elif 'VERSION_MINOR' in line:
+                minor = line.split('=')[1].strip()
+            elif 'VERSION_BUILD' in line:
+                build = line.split('=')[1].strip()
+            elif 'VERSION_ALPHA' in line:
+                alpha = line.split('=')[1].strip()
+
+            if ((major and minor and build and alpha) or
+                    '# END_VERSION_BLOCK' in line):
+                break
+    version = f"{major}.{minor}.{build}"
+    if alpha and int(alpha) > 0:
+        version += f"a{alpha}"
+    return version
+
+
+extra_files = package_files('ovos_color_parser')
+
+with open(f"{BASEDIR}/README.md", "r") as fh:
+    long_description = fh.read()
+
+setup(
+    name='ovos-color-parser',
+    version=get_version(),
+    packages=['ovos_color_parser'],
+    url='https://github.com/OpenVoiceOS/ovos-color-parser',
+    obsoletes=['ovos_color_parser'],
+    license='Apache2.0',
+    package_data={'': extra_files},
+    include_package_data=True,
+    install_requires=required('requirements.txt'),
+    author='Mycroft AI / OVOS',
+    author_email='jarbasai@mailfence.com',
+    description='OpenVoiceOS\'s multilingual text color parsing and formatting library',
+    long_description=long_description,
+    long_description_content_type="text/markdown",
+    classifiers=[
+        'Development Status :: 4 - Beta',
+        'Intended Audience :: Developers',
+        'Topic :: Text Processing :: Linguistic',
+        'License :: OSI Approved :: Apache Software License',
+
+        'Programming Language :: Python :: 3.6',
+        'Programming Language :: Python :: 3.7',
+        'Programming Language :: Python :: 3.8',
+        'Programming Language :: Python :: 3.9',
+    ]
+)
